# restaurant-management-25/docs/api.yml
openapi: 3.1.0
info:
  title: Restaurant Management API - 2025 [J]
  description: Restaurant Management API - 2025 [J]
  version: 1.0.0
servers:
  - url: 'https://localhost:8080'
paths:
  /ingredients:
    get:
      summary: Get list of ingredients
      description: List of ingredients
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
  /ingredients/{id}/prices:
    put:
      summary: Add new prices to a specific ingredient
      parameters:
        - $ref: '#/components/parameters/IngredientIdentifier'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Price'
        required: true
      responses:
        200:
          description: The specific ingredient with updated price history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
  /ingredients/{id}/stockMovements:
    put:
      summary: Add new stock movements to a specific ingredient
      parameters:
        - $ref: '#/components/parameters/IngredientIdentifier'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StockMovement'
        required: true
      responses:
        200:
          description: The specific ingredient with updated stockMovement history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
  /dishes:
    get:
      summary: Get dishes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          description: List of dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'

components:
  parameters:
    IngredientIdentifier:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        format: int32
        default: 1
    PageSize:
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
        format: int32
        default: 20
  schemas:
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        availableQuantity:
          type: integer
        actualPrice:
          type: number
          format: float
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/DishIngredient'
    DishIngredient:
      type: object
      properties:
        requiredQuantity:
          type: number
        unit:
          $ref: '#/components/schemas/Unit'
        ingredient:
          $ref: '#/components/schemas/IngredientBasicProperty'
    IngredientBasicProperty:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Ingredient:
      allOf:
        - $ref: '#/components/schemas/IngredientBasicProperty'
        - type: object
          properties:
            prices:
              type: array
              items:
                $ref: '#/components/schemas/Price'
            stockMovements:
              type: array
              items:
                $ref: '#/components/schemas/StockMovement'
            availableQuantity:
              type: number
            actualPrice:
              type: number
    Price:
      type: object
      properties:
        id:
          type: integer
          format: int64
        value:
          type: number
          format: float
        dateValue:
          type: string
          format: date
    StockMovement:
      type: object
      properties:
        id:
          type: integer
          format: int64
        quantity:
          type: number
        unit:
          $ref: '#/components/schemas/Unit'
        type:
          $ref: '#/components/schemas/StockMovementType'
        creationDatetime:
          type: string
          format: date-time
    StockMovementType:
      type: string
      enum:
        - IN
        - OUT
    Unit:
      type: string
      enum:
        - G
        - U
        - L