# restaurant-management-25/src/main/java/org/example/restaurantmanagement25/endpoint/rest/IngredientBasicProperty.java
package org.example.restaurantmanagement25.endpoint.rest;

import lombok.AllArgsConstructor;
import lombok.Getter;

@AllArgsConstructor
@Getter
public class IngredientBasicProperty {
    private Long id;
    private String name;
}