# restaurant-management-25/src/main/java/org/example/restaurantmanagement25/model/StockMovement.java
package org.example.restaurantmanagement25.model;

import lombok.*;

import java.time.Instant;

@EqualsAndHashCode
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Data
@ToString(exclude = "ingredient")
public class StockMovement {
    private Long id;
    private Ingredient ingredient;
    private Double quantity;
    private Unit unit;
    private StockMovementType movementType;
    private Instant creationDatetime;
}